style(llm): rename DrawerContext to IsInDrawerContext
