+++ apps/ledger-live-mobile/src/components/QueuedDrawer.tsx
@@ -6,7 +6,7 @@ import type { BaseModalProps } from "@ledgerhq/native-ui/components/Layout/Modal
 import { useSelector } from "react-redux";
 import { isModalLockedSelector } from "../reducers/appstate";
 import { Merge } from "../types/helpers";
-import IsInDrawerContext from "../context/DrawerContext";
+import IsInDrawerContext from "../context/IsInDrawerContext";
 
 // Purposefully removes isOpen prop so consumers can't use it directly
 export type Props = Merge<
