+++ apps/ledger-live-mobile/src/components/RequiresBLE/BluetoothDisabled.tsx
@@ -4,7 +4,7 @@ import { Icon } from "@ledgerhq/native-ui";
 import { useTranslation } from "react-i18next";
 import GenericInformationalDrawerContent from "../GenericInformationalDrawerContent";
 import GenericInformationalView from "../GenericInformationalView";
-import IsInDrawerContext from "../../context/DrawerContext";
+import IsInDrawerContext from "../../context/IsInDrawerContext";
 
 export type Props = {
   onRetry?: (() => void) | null;
