+++ apps/ledger-live-mobile/src/components/RequiresLocation/LocationDisabled.tsx
@@ -5,7 +5,7 @@ import LocationServicesDialogBox from "react-native-android-location-services-di
 import NoLocationImage from "../../icons/NoLocationImage";
 import GenericInformationalDrawerContent from "../GenericInformationalDrawerContent";
 import GenericInformationalView from "../GenericInformationalView";
-import IsInDrawerContext from "../../context/DrawerContext";
+import IsInDrawerContext from "../../context/IsInDrawerContext";
 
 type Props = {
   onRetry?: (() => void) | null;
